break out intro modal into component
